add balance to expert but have some bgs

+++ src/main/java/com/example/final_project_faz3/maktab/ir/data/model/entity/Credit.java
@@ -22,7 +22,7 @@ public class Credit {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
-    private BigDecimal balance;
+    private Long balance;
 
     @Override
     public boolean equals(Object o) {
